+++ .gitignore
@@ -1,9 +1,12 @@
 node_modules/
 package-lock.json
 config.json
-/config.dev.json
+config.dev.json
 yarn.lock
 /.idea/
-/CCConfig.json
+CCConfig.json
 /.fleet/settings.json
 */announcement.js
+
+env/**
+!env/templates/**
