--- .run/cmd.run.xml
@@ -1,8 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="cmd" type="NodeJSConfigurationType" folderName="AddCommand" singleton="false" path-to-js-file="deploy-commands.js" working-dir="$PROJECT_DIR$">
-    <envs pass-parent-envs="false">
-      <env name="NODE_ENV" value="develop" />
-    </envs>
-    <method v="2" />
-  </configuration>
-</component>
