--- .run/dev.run.xml
@@ -1,8 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="dev" type="NodeJSConfigurationType" folderName="Run" singleton="false" path-to-js-file="botmain.js" working-dir="$PROJECT_DIR$">
-    <envs pass-parent-envs="false">
-      <env name="NODE_ENV" value="development" />
-    </envs>
-    <method v="2" />
-  </configuration>
-</component>
