+++ src/botmain.js
