+++ src/functions/CCFunc.js
