+++ src/functions/TxtEasterEgg.js
