+++ src/functions/genshin.js
