+++ src/functions/logsystem.js
