+++ src/functions/statusAndMode.js
