+++ src/functions/ttGeneration.js
